[not verified] Protect: Move vul navigation to own file
